Maxent

This Class represents a MaxEnt model objects and hosts all the information related to the model.

Slots

presence

SWD. The presence locations used to train the model.

background

SWD. The backgorund locations used to train the model.

test

SWD. The test locations used to validate the model.

results

matrix. The result that usually MaxEnt provide as a csv file.

rm

numeric. The value of the regularization multiplier used to train the model.

fc

character. The feature class combination used to train the model.

iter

numeric. The number of iterations used to train the model.

type

character. The output format of the model.

lambdas

vector. The lambdas parameters of the model.

coeff

data.frame. The lambda coefficients of the model.

formula

formula. The formula used to make prediction.

lpn

numeric. Linear Predictor Normalizer.

dn

numeric. Density Normalizer.

entropy

numeric. The entropy value.

min_max

data.frame. The minimum and maximum values of the continuous variables, used for clamping.

folder

character. The path for the folder where are saved all the files produced by MaxEnt, available if the "folder" parameter is provided to the runMaxent function.