Reduce Variables

Remove variables whose contribution (permutation or percent) is less than the given threshold. The function removes one variable at time and after trains a new model to get the new variable contribution rank. If use_jk is TRUE the function checks if after removing the variable the model performance decreases (according to the given metric). In this case the function stops removing the variable even if the contribution is lower than the given threshold.

reduceVar(model, th, metric = c("auc", "tss", "aicc"), test = NULL,
  env = NULL, parallel = FALSE, use_jk = FALSE, permut = 10)

Arguments

model
SDMmodel object.

th
numeric. The contribution threshold used to remove variables.

metric
character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc", used only if use_jk is TRUE, default is "auc".

test
SWD. Test dataset used to evaluate the model, not used with aicc, and if use_jk = FALSE, default is NULL.

env
stack or brick containing the environmental variables, used only with "aicc", default is NULL.

parallel
logical, if TRUE it uses parallel computation, deafult is FALSE. Used only with AICc.

use_jk
Flag to use the Jackknife AUC test during the variable selection, if FALSE the function uses the percent variable contribution, default is FALSE.

permut
integer. Number of permutations, default is 10.

Value

The model trained using the selected variables.

Examples

# NOT RUN {
model <- reduveVar(model = model, th = 2)
# }