Tune number of Background locations

Test different sizes of background locations.

tuneBg(model, bg4test, bgs, metric = c("auc", "tss", "aicc"), env = NULL,
  parallel = FALSE, seed = NULL)

Arguments

model
Maxent object.

bg4test
SWD. The dataset with the maximum number of background locations to be tested given as MaxentSWD object.

bgs
vector. A sequence of number of background locations to test. The maximum number must be lower thatn the maximum number of background locations in bg4test. Default is the maximum number of background locations in bg4test.

metric
character. The metric used to evaluate the models, possible values are: "auc", "tss" and "aicc", default is "auc"

env
stack or brick containing the environmental variables, used only with "aicc", default is NULL.

parallel
logical, if TRUE it uses parallel computation, deafult is FALSE.

seed
integer. The value used to set the seed in order to have consistent results, default is NULL.

Value

A SDMtune object.

Details

You need package snow to use parallel computation and rgdal to save the prediction in a raster file. Parallel computation increases the speed only for big datasets due to the time necessary to create the cluster.

See also

Examples

# NOT RUN {
bg_test <- tuneBg(model, bg4test,bgs = seq(5000, 30000, 5000), metric = "auc", seed = 25)
# }